Refactor in new tqdm.py file
